修改期號長度

+++ facade/src/main/java/com/tggame/bet/vo/BetOrderVO.java
@@ -58,7 +58,7 @@ public class BetOrderVO implements java.io.Serializable {
      * 数据库字段:issue  属性显示:期号
      */
     @ApiModelProperty(value = "期号")
-    private java.lang.Integer issue;
+    private java.lang.Long issue;
     /**
      * 数据库字段:lottery_name  属性显示:彩种
      */
